Make testing more reliable and realistic
- Use `OnceLock` for `AUXILIARY_EVENT_TX` because it can't be set more than once in real air usage. This also has the benefit of removing the `static mut` rustc yells at us about.

- Move LSP client tests to integration tests. This ensures they run sequentially and in their own process. This more accurately represents real usage and avoids writing to `AUXILIARY_EVENT_TX` multiple times (or overwriting it mid test).

- Add `TESTING` global to turn off logs during testing. Emitting log notifications to the Client during testing is problematic because we use `recv_response()` to track request/response lifecycle and log requests end up getting in the way. Ideally I think we'd find a way to filter out the logs and keep "recv()"-ing instead of using a global like this to avoid them entirely.
