+++ crates/lsp/src/handlers_format.rs
@@ -40,51 +40,3 @@ pub(crate) fn document_formatting(
     let edits = to_proto::replace_all_edit(&doc.line_index, &doc.contents, &output)?;
     Ok(Some(edits))
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::{
-        documents::Document, tower_lsp::init_test_client, tower_lsp_test_client::TestClientExt,
-    };
-
-    #[tests_macros::lsp_test]
-    async fn test_format() {
-        let mut client = init_test_client().await;
-
-        #[rustfmt::skip]
-        let doc = Document::doodle(
-"
-1
-2+2
-3 + 3 +
-3",
-        );
-
-        let formatted = client.format_document(&doc).await;
-        insta::assert_snapshot!(formatted);
-
-        client
-    }
-
-    // https://github.com/posit-dev/air/issues/61
-    #[tests_macros::lsp_test]
-    async fn test_format_minimal_diff() {
-        let mut client = init_test_client().await;
-
-        #[rustfmt::skip]
-        let doc = Document::doodle(
-"1
-2+2
-3
-",
-        );
-
-        let edits = client.format_document_edits(&doc).await.unwrap();
-        assert!(edits.len() == 1);
-
-        let edit = &edits[0];
-        assert_eq!(edit.new_text, " + ");
-
-        client
-    }
-}
