+++ crates/lsp/tests/fixtures/mod.rs
@@ -1,0 +1,5 @@
+// For some reason this setup gives false alarms about dead code
+#[allow(dead_code)]
+pub mod test_client;
+#[allow(dead_code)]
+pub mod tower_lsp_test_client;
